requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy==2.0.36
pydantic==2.9.2
PyMySQL==1.1.1
python-dotenv==1.0.1